% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PML_helper.R
\name{add_ETsum}
\alias{add_ETsum}
\title{calculate ETsum}
\usage{
add_ETsum(l)
}
\arguments{
\item{l}{List object, with the elements at least of \code{Ecr}, \code{Eca}, \code{Ei}, \code{Es}.}
}
\description{
ET_sum = Ec + Ei + Es$
}
\keyword{internal}
