% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\alias{options_param}
\alias{d_CO2}
\title{global variables}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18 rows and 14 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1932 rows and 9 columns.
}
\usage{
options_param

d_CO2
}
\description{
\itemize{
\item \code{options_param}: PMLV2 parameter of version 017
}

\itemize{
\item \code{d_CO2}: CO2 concentration in ppm
}
}
\keyword{datasets}
