% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R
\name{ET_helper}
\alias{ET_helper}
\alias{get_lambda}
\alias{get_Rn}
\alias{get_VPD}
\title{ET helper}
\usage{
get_lambda(Tavg)

get_Rn(Rs, Rln, Tavg, albedo, emiss)

get_VPD(Pa, q, Tmax, Tmin)
}
\arguments{
\item{Tavg}{daily mean temperature, deg}

\item{Rs}{shortwave inward solar radiation, W m-2}

\item{Rln}{longwave inward solar radiation, W m-2}

\item{albedo}{shortwave albedo, -}

\item{emiss}{longwave emissivity, -}

\item{Pa}{air pressure, kPa}

\item{q}{specific humidity, kg kg-1}

\item{Tmax, Tmin}{maximum and minimum temperature, deg C}
}
\description{
\itemize{
\item \code{get_lambda}: latent heat of vaporization (~2500 J g-1)
\item \code{get_Rn}: net solar radiation (W m-2)
\item \code{get_VPD}: vapor pressure deficit (kPa)
}
}
