% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R
\name{get_lambda}
\alias{get_lambda}
\title{latent heat of vaporization}
\usage{
get_lambda(Tavg)
}
\arguments{
\item{Tavg}{daily mean temperature, deg}
}
\value{
\code{lambda}: J g-1
}
\description{
latent heat of vaporization
}
