% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R
\name{get_Rn}
\alias{get_Rn}
\title{Get net solar radiation}
\usage{
get_Rn(Rs, Rln, Tavg, albedo, emiss)
}
\arguments{
\item{Rs}{shortwave inward solar radiation, W m-2}

\item{Rln}{longwave inward solar radiation, W m-2}

\item{Tavg}{daily mean temperature, deg}

\item{albedo}{shortwave albedo, -}

\item{emiss}{longwave emissivity, -}
}
\value{
\code{Rn}: net solar radiation, W m-2
}
\description{
Get net solar radiation
}
