% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R
\name{get_VPD}
\alias{get_VPD}
\title{vapor pressure deficit}
\usage{
get_VPD(Pa, q, Tmax, Tmin)
}
\arguments{
\item{Pa}{air pressure, kPa}

\item{q}{specific humidity, kg kg-1}

\item{Tmax, Tmin}{maximum and minimum temperature, deg C}
}
\value{
\code{VPD}: kPa
}
\description{
vapor pressure deficit
}
