% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PML_parameters.R
\docType{data}
\name{param0_PML}
\alias{param0_PML}
\title{PML parameter boundaries}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 5 columns.
}
\usage{
param0_PML
}
\description{
PML parameter boundaries
}
\keyword{datasets}
