% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_2GEE.R
\name{parameters_2GEE}
\alias{parameters_2GEE}
\alias{parameters_2R}
\title{Transform parameters to GEE}
\usage{
parameters_2GEE(file)

parameters_2R(file)
}
\arguments{
\item{file}{The file path of parameter, in the csv format}
}
\description{
parameter transfer scheme:
\itemize{
\item DNF(4) = EBF(2)
\item CSH(7) = OSH(8)
\item CNV(15) = CRO(13)
\item \code{14} (Urban and Built-Up)           = 6  (mixed forest)
\item \code{17} (Barren or Sparsely Vegetated) = 11 (grassland)
}
}
\examples{
\dontrun{
parameters_2GEE("data-raw/PMLV2_parameters_20181202_110828.csv")
}
}
