% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PML.R
\name{PML}
\alias{PML}
\title{Calculation of actual evaporation and GPP using the PML_V2 for global land
surface for each time step (16-day).}
\usage{
PML(
  data,
  par = NULL,
  V2 = TRUE,
  CO2 = NULL,
  ratio = 1,
  return.Es = TRUE,
  IGBPcode = NULL,
  scale = c("site", "regional")
)
}
\arguments{
\item{data}{A data.frame with the columns of
\itemize{
\item \code{Prcp}   : Precipitation, mm
\item \code{Tavg}   : daily average temperature, deg C
\item \code{Rs}     : shortwave radiation, W m-2
\item \code{Pa}     : surface air pressure, kPa
\item \code{U2}     : wind speed, m/s
\item \code{LAI}    : m2 m-2
}

(can be calculated by other variables)
\itemize{
\item \code{Rn}     : net solar radiation, W m-2
\itemize{
\item \code{albedo}
\item \code{emiss}
\item \code{Rln}, W m-2 (plus Rs, Tavg): \code{Rn = get_Rn(Rs, Rln, Tavg, albedo, emiss)}
}
\item \code{VPD}    : Vapor pressure deficit, kPa
\itemize{
\item \code{q}, kg/kg (plus Pa): \code{VPD = get_VPD(Pa, q)}
}
}

(optional)
\itemize{
\item \code{CO2}     : about 380-570, umol mol-1
\item \code{IGBPname}: If not specified parameters, this column is necessary for matching
parameter.
}

(optional, for speeding up optimization)
\itemize{
\item \code{lambda}  :
\item \code{rou_a}   :
\item \code{gama}    :
\item \code{epsilon} :
}}

\item{par}{PML_V2 parameters
\itemize{
\item \code{hc}      : canopy height (m)
\item \code{Alpha}   : initial photochemical efficiency, 0.02-0.08
\item \code{m}       : the initla slope of CO2 response curve \verb{[umol m-2 s-1]/[umol mol-1]}
\item \code{Am_25}   : 10 - 150
\item \code{D0}      : the value of VPD when stomtal conductance is reduced
\item \code{kQ}      : 0.7,
\item \code{kA}      : 0.7, light extinction coefficient
\item \code{S_sls}   :
\item \code{fER0}    :
\item \code{VPDmin}  :
\item \code{VPDmax}  :
}}

\item{V2}{Run PML_V2? If false, PML_V1 will be used.}

\item{CO2}{A constant number. If provided, CO2 in the \code{data} will be replaced.}

\item{ratio}{land cover ratio}

\item{return.Es}{Because \code{fval_soil} need moving average in temporal. We can't
get it directly if the \code{data} is Spatial pixels.}

\item{IGBPname}{IGBPname (not used)}
}
\value{
\itemize{
\item \code{GPP}: gC m-2 d-1, Gross primary product
\item \code{Ec} : mm d-1, Vegetation transpiration
\item \code{Es} : mm d-1, Soil evaporation |
\item \code{Ei} : mm d-1, Interception from vegetation canopy
\item \code{Es_eq}: mm d-1, Soil evaporation at equilibrium
\item \code{Pi}: mm d-1, \code{Prcp - Ei}
}
}
\description{
This algorithm fits to high resolution data data in GEE.
}
\references{
\enumerate{
\item Zhang, Y., Kong, D., Gan, R., Chiew, F. H. S., McVicar, T. R., Zhang, Q.,
& Yang, Y. (2019). Coupled estimation of 500 m and 8-day resolution global
evapotranspiration and gross primary production in 2002–2017. Remote
Sensing of Environment, 222, 165–182. \url{doi:10.1016/j.rse.2018.12.031}.
\item Kong, D., Zhang, Y., Gu, X., & Wang, D. (2019). A robust method for
reconstructing global MODIS EVI time series on the Google Earth Engine.
ISPRS Journal of Photogrammetry and Remote Sensing, 155, 13–24.
\url{https://doi.org/10.1016/j.isprsjprs.2019.06.014}
\item basic units change, \url{http://www.fao.org/docrep/X0490E/x0490e0i.htm}
}
}
