% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_calib.R
\name{PML_calib}
\alias{PML_calib}
\alias{PML_predict}
\alias{PML_gof}
\alias{PML_goal}
\title{Calibration for \code{PML} model}
\usage{
PML_calib(data, IGBPcode = 1, of_gof = NSE, ..., verbose = TRUE, maxn = 1000)

PML_predict(data, par, IGBPcode = NULL)

PML_gof(dat, of_gof = NSE, detailed = TRUE)

PML_goal(par, data, IGBPcode = NULL, ..., of_gof = NSE, detailed = FALSE)
}
\arguments{
\item{data}{A data.frame with the columns of
\itemize{
\item \code{Prcp}   : Precipitation, mm
\item \code{Tavg}   : daily average temperature, deg C
\item \code{Rs}     : shortwave radiation, W m-2
\item \code{Pa}     : surface air pressure, kPa
\item \code{U2}     : wind speed, m/s
\item \code{LAI}    : m2 m-2
}

(can be calculated by other variables)
\itemize{
\item \code{Rn}     : net solar radiation, W m-2
\itemize{
\item \code{albedo}
\item \code{emiss}
\item \code{Rln}, W m-2 (plus Rs, Tavg): \code{Rn = get_Rn(Rs, Rln, Tavg, albedo, emiss)}
}
\item \code{VPD}    : Vapor pressure deficit, kPa
\itemize{
\item \code{q}, kg/kg (plus Pa): \code{VPD = get_VPD(Pa, q)}
}
}

(optional)
\itemize{
\item \code{CO2}     : about 380-570, umol mol-1
\item \code{IGBPname}: If not specified parameters, this column is necessary for matching
parameter.
}

(optional, for speeding up optimization)
\itemize{
\item \code{lambda}  :
\item \code{rho_a}   : air density, \verb{[g m-3]}
\item \code{gama}    :
\item \code{epsilon} :
}}

\item{IGBPcode}{(optional) \link{IGBP_006} code. If \code{hc_raw} specified in \code{par}, \code{IGBPcode} can be ignored.
Otherwise, \code{IGBPcode} should be provided.}

\item{...}{ignored}

\item{par}{PML_V2 parameters (see Zhang 2019 for details).
\itemize{
\item \code{hc}      : canopy height (m)
\item \code{Alpha}   : initial photochemical efficiency, 0.02-0.08
\item \code{m}       : the initla slope of CO2 response curve \verb{[umol m-2 s-1]/[umol mol-1]}
\item \code{Am_25}   : 10 - 150
\item \code{D0}      : the value of VPD when stomtal conductance is reduced
\item \code{kQ}      : 0.7,
\item \code{kA}      : 0.7, light extinction coefficient
\item \code{S_sls}   :
\item \code{fER0}    :
\item \code{VPDmin}  :
\item \code{VPDmax}  :
}}
}
\description{
Calibration for \code{PML} model
}
