% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PML_region.R
\name{PML_region}
\alias{PML_region}
\title{PML_region}
\usage{
PML_region(
  f_mete,
  f_lc,
  outfile = NULL,
  poly = NULL,
  ...,
  overwrite = TRUE,
  verbose = FALSE
)
}
\description{
Running one year each time.
}
