% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PML.R
\name{PML_years}
\alias{PML_years}
\title{PML V2 dynamic and static experiments}
\usage{
PML_years(
  files_forcing,
  lst_LC_ratio,
  type = c("dynamic", "static")[2],
  is_save = TRUE,
  overwrite = TRUE,
  outdir = "OUTPUT/static",
  prefix = "PML_V2-RAW-static-8day-TP_010deg"
)
}
\description{
PML V2 dynamic and static experiments
}
