% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMLV2-package.R
\docType{package}
\name{PMLV2-package}
\alias{PMLV2}
\alias{PMLV2-package}
\title{PMLV2: Penman-Monteith-Leuning Evapotranspiration model V2}
\description{
Simulate evapotranspiration (ET), its three components, and gross primary product (GPP). This model performs well against observations at 95 flux sites across the globe. They are similar to or noticeably better than major state-of-the-art ET and GPP products widely used by water and ecology science communities. Reference: Zhang, Y., Kong, D., Gan, R., Chiew, F. H. S., McVicar, T. R., Zhang, Q., & Yang, Y. (2019). Coupled estimation of 500 m and 8-day resolution global evapotranspiration and gross primary production in 2002–2017. Remote Sensing of Environment, 222, 165–182. \doi{10.1016/j.rse.2018.12.031}.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
