% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_forcing_flux.R
\name{tidy_forcing_flux}
\alias{tidy_forcing_flux}
\title{Tidy meteorological forcing of FLUXNET}
\usage{
tidy_forcing_flux(df)
}
\arguments{
\item{df}{A data.frame read from FLUXNET
\itemize{
\item \code{Pa}     : kPa
\item \code{Rs}     : W m-2
\item \code{Tavg}   : degC
\item \code{Albedo} : -
\item \code{Emiss}  : -
\item \code{LAI}    : -
\item \code{LE}     : W m-2, (will be converted to mm/d)
\item \code{GPP_DT},\code{GPP_NT}: \code{GPPobs = nanmean(GPP_DT, GPP_NT)}, \verb{gC m-2 d-1}
}}
}
\value{
\itemize{
\item lambda
\item Rn
\item Eeq
\item rho_a
\item gama
\item epsilon
\item fval_soil
}
}
\description{
Tidy meteorological forcing of FLUXNET
}
\references{
\enumerate{
\item Convert PAR from MJ/m2/d to umol m-2 s-1, \url{http://www.egc.com/useful_info_lighting.php}
\item \url{https://appgeodb.nancy.inra.fr/biljou/pdf/Allen_FAO1998.pdf}
}
}
