% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PML.R
\name{tidy_forcing_yearly}
\alias{tidy_forcing_yearly}
\title{tidy_forcing_yearly}
\usage{
tidy_forcing_yearly(file, l_static = NULL, write2nc = FALSE, overwrite = FALSE)
}
\arguments{
\item{l_static}{NULL}
}
\description{
tidy_forcing_yearly
}
